# clientv3/concurrency: Mutex blocking