--- a/bill-of-materials.json
+++ b/bill-of-materials.json
@@ -378,6 +378,15 @@
 		]
 	},
 	{
+		"project": "golang.org/x/sync/semaphore",
+		"licenses": [
+			{
+				"type": "BSD 3-clause \"New\" or \"Revised\" License",
+				"confidence": 0.9663865546218487
+			}
+		]
+	},
+	{
 		"project": "golang.org/x/sys/unix",
 		"licenses": [
 			{